+++ panelapp/pa_locus_list_api_tests.py
@@ -1,12 +1,16 @@
 import json
+from collections import defaultdict
+
 import mock
 import responses
+import tenacity
 from django.core.management import call_command, CommandError
 from django.urls.base import reverse
+from requests import Response
+from urllib3.exceptions import MaxRetryError
 
-from seqr.models import LocusList
-from seqr.views.apis.locus_list_api import locus_lists, locus_list_info, add_project_locus_lists, \
-    delete_project_locus_lists
+from panelapp.panelapp_utils import _get_all_genes
+from seqr.views.apis.locus_list_api import locus_lists, locus_list_info
 from seqr.views.apis.locus_list_api_tests import BaseLocusListAPITest
 from seqr.views.utils.test_utils import AuthenticationTestCase, LOCUS_LIST_FIELDS
 
@@ -55,22 +59,21 @@ class PaLocusListAPITest(AuthenticationTestCase, BaseLocusListAPITest):
         # Given all PanelApp gene lists and associated genes
         au_panels_p1_url = '{}/panels/?page=1'.format(PANEL_APP_API_URL_AU)
         au_panels_p2_url = '{}/panels/?page=2'.format(PANEL_APP_API_URL_AU)
-        uk_panels_p1_url = '{}/panels/?page=1'.format(PANEL_APP_API_URL_UK)
-        au_genes_260_url = '{}/panels/{}/genes/?page=1'.format(PANEL_APP_API_URL_AU, 260)
-        au_genes_3069_url = '{}/panels/{}/genes/?page=1'.format(PANEL_APP_API_URL_AU, 3069)
-        uk_genes_260_url = '{}/panels/{}/genes/?page=1'.format(PANEL_APP_API_URL_UK, 260)
+        au_genes_url = '{}/genes/?page=1'.format(PANEL_APP_API_URL_AU)
         au_panels_p1_json = _get_json_from_file('panelapp/test_resources/au_panelapp_panels_p1.json')
         au_panels_p2_json = _get_json_from_file('panelapp/test_resources/au_panelapp_panels_p2.json')
+        au_genes_json = _get_json_from_file('panelapp/test_resources/au_panelapp_genes.json')
+
+        uk_panels_p1_url = '{}/panels/?page=1'.format(PANEL_APP_API_URL_UK)
+        uk_genes_url = '{}/genes/?page=1'.format(PANEL_APP_API_URL_UK)
         uk_panels_p1_json = _get_json_from_file('panelapp/test_resources/uk_panelapp_panels_p1.json')
-        au_genes_260_json = _get_json_from_file('panelapp/test_resources/au_panel_260_genes.json')
-        au_genes_3069_json = _get_json_from_file('panelapp/test_resources/au_panel_3069_genes.json')
-        uk_genes_260_json = _get_json_from_file('panelapp/test_resources/uk_panel_260_genes.json')
+        uk_genes_json = _get_json_from_file('panelapp/test_resources/uk_panelapp_genes.json')
+
         responses.add(responses.GET, au_panels_p1_url, json=au_panels_p1_json, status=200)
         responses.add(responses.GET, au_panels_p2_url, json=au_panels_p2_json, status=200)
+        responses.add(responses.GET, au_genes_url, json=au_genes_json, status=200)
         responses.add(responses.GET, uk_panels_p1_url, json=uk_panels_p1_json, status=200)
-        responses.add(responses.GET, au_genes_260_url, json=au_genes_260_json, status=200)
-        responses.add(responses.GET, au_genes_3069_url, json=au_genes_3069_json, status=200)
-        responses.add(responses.GET, uk_genes_260_url, json=uk_genes_260_json, status=200)
+        responses.add(responses.GET, uk_genes_url, json=uk_genes_json, status=200)
 
         # URl argument is required
         with self.assertRaises(CommandError) as err:
@@ -168,3 +171,29 @@ class PaLocusListAPITest(AuthenticationTestCase, BaseLocusListAPITest):
         self.assertEqual(response.status_code, 200)
         locus_lists_dict = response.json()['locusListsByGuid']
         self.assertSetEqual(set(locus_lists_dict.keys()), {LOCUS_LIST_GUID})
+
+    @mock.patch("panelapp.panelapp_utils.requests.get")
+    def test_get_all_genes_exhausts_retries(self, mock_get_request):
+        url = '{}/genes/?page=1'.format(PANEL_APP_API_URL_UK)
+        request_error = MaxRetryError(pool=mock.MagicMock(), url=url)
+        mock_get_request.side_effect = [request_error] * 5
+        with self.assertRaises(tenacity.RetryError):
+            _get_all_genes(url, defaultdict(list))
+
+    @mock.patch("panelapp.panelapp_utils.requests.get")
+    def test_get_all_genes_retries_success(self, mock_get_request):
+        url = '{}/genes/?page=1'.format(PANEL_APP_API_URL_UK)
+        request_error = MaxRetryError(pool=mock.MagicMock(), url=url)
+        page_1 = Response()
+        page_1.status_code = 200
+        page_1._content = (b'{"next":"https://test-panelapp.url.uk/api/v1/genes/?page=2","results": [{"panel":'
+                           b'{"id": 1207, "name": "Acute intermittent porphyria"}}]}')
+        page_2 = Response()
+        page_2.status_code = 200
+        page_2._content = b'{"results": [{"panel": {"id": 1141, "name": "Acute rhabdomyolysis"}}]}'
+        mock_get_request.side_effect = [request_error] * 4 + [page_1] + [request_error] * 4 + [page_2]
+        expected_res = {
+            1207: [{'panel': {'id': 1207, 'name': 'Acute intermittent porphyria'}}],
+            1141: [{'panel': {'id': 1141, 'name': 'Acute rhabdomyolysis'}}],
+        }
+        self.assertEqual(_get_all_genes(url, defaultdict(list)), expected_res)