+++ ui/pages/SummaryData/components/IndividualMetadata.jsx
@@ -1,8 +1,12 @@
-import React from 'react'
+import { connect } from 'react-redux'
 
+import { getUser } from 'redux/selectors'
 import { BaseSemanticInput, BooleanCheckbox } from 'shared/components/form/Inputs'
-import { FAMILY_ANALYSIS_STATUS_LOOKUP, VARIANT_METADATA_COLUMNS } from 'shared/utils/constants'
-import LoadReportTable from './LoadReportTable'
+import LoadReportTable from 'shared/components/table/LoadReportTable'
+import { VARIANT_METADATA_COLUMNS, BASE_FAMILY_METADATA_COLUMNS } from 'shared/utils/constants'
+
+const ALL_PROJECTS_PATH = 'all'
+const GREGOR_PROJECT_PATH = 'gregor'
 
 const FIELDS = [
   {
@@ -27,29 +31,16 @@ const AIRTABLE_FIELDS = [
 
 const CORE_COLUMNS = [
   { name: 'participant_id', secondaryExportColumn: 'individual_guid' },
-  { name: 'pmid_id' },
   { name: 'paternal_id', secondaryExportColumn: 'paternal_guid' },
   { name: 'maternal_id', secondaryExportColumn: 'maternal_guid' },
   { name: 'proband_relationship' },
   { name: 'sex' },
   { name: 'ancestry' },
-  { name: 'condition_id' },
-  { name: 'known_condition_name', secondaryExportColumn: 'disorders' },
   { name: 'affected_status' },
   { name: 'hpo_present', style: { minWidth: '400px' } },
   { name: 'hpo_absent', style: { minWidth: '400px' } },
-  { name: 'phenotype_description', style: { minWidth: '200px' } },
-  { name: 'analysis_groups' },
-  {
-    name: 'analysisStatus',
-    content: 'analysis_status',
-    format: ({ analysisStatus }) => FAMILY_ANALYSIS_STATUS_LOOKUP[analysisStatus]?.name,
-  },
-  { name: 'solve_status' },
   { name: 'MME' },
-  { name: 'data_type' },
-  { name: 'date_data_generation', secondaryExportColumn: 'filter_flags' },
-  { name: 'consanguinity' },
+  ...BASE_FAMILY_METADATA_COLUMNS,
   { name: 'family_history' },
 ]
 
@@ -63,30 +54,37 @@ const AIRTABLE_COLUMNS = [
   { name: 'sample_provider' },
 ]
 
+const ANALYST_VIEW_ALL_PAGES = [
+  { name: 'GREGoR', downloadName: 'All_GREGoR_Projects', path: GREGOR_PROJECT_PATH },
+  { name: 'Broad', downloadName: 'All_AnVIL_Projects', path: ALL_PROJECTS_PATH },
+]
+const VIEW_ALL_PAGES = [{ name: 'my', downloadName: 'All_Projects', path: ALL_PROJECTS_PATH }]
+
 const getColumns = (data) => {
   const maxSavedVariants = Math.max(1, ...(data || []).map(row => row.num_saved_variants))
   const hasAirtable = data && data[0] && data[0][AIRTABLE_DBGAP_SUBMISSION_FIELD]
   return [...CORE_COLUMNS, ...(hasAirtable ? AIRTABLE_COLUMNS : [])].concat(
     ...[...Array(maxSavedVariants).keys()].map(i => VARIANT_METADATA_COLUMNS.map(
-      ({ name, format, fieldName, ...col }) => ({
+      ({ name, format, secondaryExportColumn, ...col }) => ({
         name: `${name}-${i + 1}`,
-        secondaryExportColumn: name === 'gene' ? `gene_id-${i + 1}` : null,
-        format: format ? row => format({ [fieldName]: row[`${fieldName}-${i + 1}`] }) : null,
+        secondaryExportColumn: secondaryExportColumn && `${secondaryExportColumn}-${i + 1}`,
+        format: format ? row => format({ [name]: row[`${name}-${i + 1}`] }) : null,
         ...col,
       }),
     )),
   )
 }
 
-const IndividualMetadata = props => (
-  <LoadReportTable
-    getColumns={getColumns}
-    allQueryFields={AIRTABLE_FIELDS}
-    queryFields={FIELDS}
-    urlPath="individual_metadata"
-    idField="participant_id"
-    {...props}
-  />
-)
+const mapStateToProps = (state, ownProps) => {
+  const user = getUser(state)
+  return {
+    getColumns,
+    queryFields: (user.isAnalyst && ownProps.match.params.projectGuid !== ALL_PROJECTS_PATH) ? AIRTABLE_FIELDS : FIELDS,
+    viewAllPages: (user.isAnalyst ? ANALYST_VIEW_ALL_PAGES : VIEW_ALL_PAGES),
+    urlBase: 'summary_data/individual_metadata',
+    idField: 'participant_id',
+    fileName: 'Metadata',
+  }
+}
 
-export default IndividualMetadata
+export default connect(mapStateToProps)(LoadReportTable)